refactor: DifficultyGraph components/ 폴더 아래로 이동
